Rectangles count was added in WPF
